requirements.txt
Pillow==9.1.1
numpy==1.22.4
scikit-learn==1.1.1